+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/cards/AbstractCard/RenderDescriptionEnergy.java
@@ -8,6 +8,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.evacipated.cardcrawl.modthespire.lib.*;
 import com.evacipated.cardcrawl.modthespire.patcher.PatchingException;
 import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.DescriptionLine;
 import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.helpers.FontHelper;
 import javassist.CannotCompileException;
@@ -42,7 +43,30 @@ public class RenderDescriptionEnergy
 
     @SpirePatch(
             clz=AbstractCard.class,
+            method="renderDescriptionCN"
+    )
+    public static class AlterTmpCN
+    {
+        @SpireInsertPatch(
+                rloc=35,
+                localvars={"tmp"}
+        )
+        public static void Insert(AbstractCard __instance, SpriteBatch sb, @ByRef String[] tmp)
+        {
+            Matcher m = r.matcher(tmp[0]);
+            if (m.find()) {
+                tmp[0] = "[E]" + (m.group(2).equals(".") ? "." : "") + " ";
+            }
+        }
+    }
+
+    @SpirePatch(
+            clz=AbstractCard.class,
             method="renderDescription"
+    )
+    @SpirePatch(
+            clz=AbstractCard.class,
+            method="renderDescriptionCN"
     )
     public static class RenderSmallEnergyOrb
     {
@@ -56,13 +80,13 @@ public class RenderDescriptionEnergy
                                   BitmapFont font, Color textColor, @ByRef String[] tmp, GlyphLayout gl)
         {
             Matcher m = r.matcher(tmp[0]);
-            if (m.find()) {
+            if (tmp[0].equals("[E]") || m.find()) {
                 gl.width = CARD_ENERGY_IMG_WIDTH * __instance.drawScale;
                 float tmp2 = (__instance.description.size() - 4) * spacing;
                 __instance.renderSmallEnergy(sb, BaseMod.getCardSmallEnergy(__instance),
                         (start_x[0] - __instance.current_x) / Settings.scale / __instance.drawScale,
                         -100.0f - ((__instance.description.size() - 4.0f) / 2.0f - i + 1.0f) * spacing);
-                if (m.group(2).equals(".")) {
+                if (!tmp[0].equals("[E]") && m.group(2).equals(".")) {
                     FontHelper.renderRotatedText(sb, font, ".",
                             __instance.current_x, __instance.current_y,
                             start_x[0] - __instance.current_x + CARD_ENERGY_IMG_WIDTH * __instance.drawScale,
@@ -88,7 +112,52 @@ public class RenderDescriptionEnergy
 
     @SpirePatch(
             clz=AbstractCard.class,
+            method="initializeDescriptionCN"
+    )
+    public static class FixEForChinese
+    {
+        @SpireInsertPatch(
+                locator=Locator.class,
+                localvars={"word", "currentWidth", "currentLine", "numLines", "CARD_ENERGY_IMG_WIDTH", "CN_DESC_BOX_WIDTH"}
+        )
+        public static void Insert(AbstractCard __instance, @ByRef String[] word, @ByRef float[] currentWidth,
+                                  @ByRef StringBuilder[] currentLine, @ByRef int[] numLines,
+                                  float CARD_ENERGY_IMG_WIDTH, float CN_DESC_BOX_WIDTH)
+        {
+            if (word[0].equals("[E]")) {
+                if (currentWidth[0] + CARD_ENERGY_IMG_WIDTH > CN_DESC_BOX_WIDTH) {
+                    ++numLines[0];
+                    __instance.description.add(new DescriptionLine(currentLine[0].toString(), currentWidth[0]));
+                    currentLine[0] = new StringBuilder();
+                    currentWidth[0] = CARD_ENERGY_IMG_WIDTH;
+                    currentLine[0].append(" ").append(word[0]).append(" ");
+                } else {
+                    currentLine[0].append(" ").append(word[0]).append(" ");
+                    currentWidth[0] += CARD_ENERGY_IMG_WIDTH;
+                }
+                word[0] = "";
+            }
+        }
+
+        private static class Locator extends SpireInsertLocator
+        {
+            public int[] Locate(CtBehavior ctMethodToPatch) throws CannotCompileException, PatchingException
+            {
+                com.evacipated.cardcrawl.modthespire.lib.Matcher finalMatcher = new com.evacipated.cardcrawl.modthespire.lib.Matcher.MethodCallMatcher(
+                        String.class, "toCharArray");
+
+                return LineFinder.findInOrder(ctMethodToPatch, finalMatcher);
+            }
+        }
+    }
+
+    @SpirePatch(
+            clz=AbstractCard.class,
             method="initializeDescription"
+    )
+    @SpirePatch(
+            clz=AbstractCard.class,
+            method="initializeDescriptionCN"
     )
     public static class AlterEnergyKeyword
     {
@@ -96,9 +165,9 @@ public class RenderDescriptionEnergy
                 locator=Locator.class,
                 localvars={"word"}
         )
-        public static void Insert(AbstractCard __instance, String word)
+        public static void Insert(AbstractCard __instance, @ByRef String[] word)
         {
-            if (word.equals("[E]") && !__instance.keywords.contains("[E]")) {
+            if (word[0].equals("[E]") && !__instance.keywords.contains("[E]")) {
                 __instance.keywords.add("[E]");
             }
         }
@@ -108,9 +177,9 @@ public class RenderDescriptionEnergy
             public int[] Locate(CtBehavior ctMethodToPatch) throws CannotCompileException, PatchingException
             {
                 com.evacipated.cardcrawl.modthespire.lib.Matcher finalMatcher = new com.evacipated.cardcrawl.modthespire.lib.Matcher.MethodCallMatcher(
-                        "java.lang.String", "toLowerCase");
+                        String.class, "toLowerCase");
 
-                return LineFinder.findInOrder(ctMethodToPatch, new ArrayList<com.evacipated.cardcrawl.modthespire.lib.Matcher>(), finalMatcher);
+                return LineFinder.findInOrder(ctMethodToPatch, finalMatcher);
             }
         }