+++ CHANGELOG.md
@@ -514,3 +514,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Allow Dynamic Variables to show in smith view upgraded (kiooeht)
 * Maybe make card description patches work for Chinese:
   * Shrink long descriptions (kiooeht)
+  * [E] in small card descriptions (kiooeht)