+++ CHANGELOG.md
@@ -515,3 +515,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Maybe make card description patches work for Chinese:
   * Shrink long descriptions (kiooeht)
   * [E] in small card descriptions (kiooeht)
+  * [E] in SingleCardViewPopup (kiooeht)