+++ CHANGELOG.md
@@ -516,3 +516,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
   * Shrink long descriptions (kiooeht)
   * [E] in small card descriptions (kiooeht)
   * [E] in SingleCardViewPopup (kiooeht)
+  * Custom Dynamic variables (kiooeht)